move to duckdb.experimental.spark | also fix issue in 'test_spark_group_by.py'
