--- tools/pythonpkg/duckdb/spark/__init__.py
@@ -1,7 +1,0 @@
-from .sql import SparkSession, DataFrame
-from .conf import SparkConf
-from .context import SparkContext
-from ._globals import _NoValue
-from .exception import ContributionsAcceptedError
-
-__all__ = ["SparkSession", "DataFrame", "SparkConf", "SparkContext", "ContributionsAcceptedError"]
