--- tools/pythonpkg/duckdb/spark/exception.py
@@ -1,9 +1,0 @@
-class ContributionsAcceptedError(NotImplementedError):
-    """
-    This method is not planned to be implemented, if you would like to implement this method
-    or show your interest in this method to other members of the community,
-    feel free to open up a PR or a Discussion over on https://github.com/duckdb/duckdb
-    """
-
-
-__all__ = ["ContributionsAcceptedError"]
