--- tools/pythonpkg/duckdb/spark/sql/__init__.py
@@ -1,7 +1,0 @@
-from .session import SparkSession
-from .readwriter import DataFrameWriter
-from .dataframe import DataFrame
-from .conf import RuntimeConfig
-from .catalog import Catalog
-
-__all__ = ["SparkSession", "DataFrame", "RuntimeConfig", "DataFrameWriter", "Catalog"]
