--- tools/pythonpkg/duckdb/spark/sql/functions.py
@@ -1,388 +1,0 @@
-from .column import Column, _get_expr
-from typing import Any, Callable, overload, Union
-
-from duckdb import CaseExpression, ConstantExpression, ColumnExpression, FunctionExpression, Expression
-from ._typing import ColumnOrName
-from ..exception import ContributionsAcceptedError
-
-
-def _invoke_function_over_columns(name: str, *cols: "ColumnOrName") -> Column:
-    """
-    Invokes n-ary JVM function identified by name
-    and wraps the result with :class:`~pyspark.sql.Column`.
-    """
-    cols = [_to_column(expr) for expr in cols]
-    return _invoke_function(name, *cols)
-
-
-def col(column: str):
-    return Column(ColumnExpression(column))
-
-
-def upper(col: "ColumnOrName") -> Column:
-    """
-    Converts a string expression to upper case.
-
-    .. versionadded:: 1.5.0
-
-    .. versionchanged:: 3.4.0
-        Supports Spark Connect.
-
-    Parameters
-    ----------
-    col : :class:`~pyspark.sql.Column` or str
-        target column to work on.
-
-    Returns
-    -------
-    :class:`~pyspark.sql.Column`
-        upper case values.
-
-    Examples
-    --------
-    >>> df = spark.createDataFrame(["Spark", "PySpark", "Pandas API"], "STRING")
-    >>> df.select(upper("value")).show()
-    +------------+
-    |upper(value)|
-    +------------+
-    |       SPARK|
-    |     PYSPARK|
-    |  PANDAS API|
-    +------------+
-    """
-    return _invoke_function_over_columns("upper", col)
-
-
-def when(condition: "Column", value: Any) -> Column:
-    if not isinstance(condition, Column):
-        raise TypeError("condition should be a Column")
-    v = _get_expr(value)
-    expr = CaseExpression(condition.expr, v)
-    return Column(expr)
-
-
-def _inner_expr_or_val(val):
-    return val.expr if isinstance(val, Column) else val
-
-
-def struct(*cols: Column) -> Column:
-    return Column(FunctionExpression('struct_pack', *[_inner_expr_or_val(x) for x in cols]))
-
-
-def lit(col: Any) -> Column:
-    return col if isinstance(col, Column) else Column(ConstantExpression(col))
-
-
-def _invoke_function(function: str, *arguments):
-    return Column(FunctionExpression(function, *arguments))
-
-
-def _to_column(col: ColumnOrName) -> Expression:
-    return col.expr if isinstance(col, Column) else ColumnExpression(col)
-
-
-def regexp_replace(str: "ColumnOrName", pattern: str, replacement: str) -> Column:
-    r"""Replace all substrings of the specified string value that match regexp with rep.
-
-    .. versionadded:: 1.5.0
-
-    Examples
-    --------
-    >>> df = spark.createDataFrame([('100-200',)], ['str'])
-    >>> df.select(regexp_replace('str', r'(\d+)', '--').alias('d')).collect()
-    [Row(d='-----')]
-    """
-    return _invoke_function(
-        "regexp_replace", _to_column(str), ConstantExpression(pattern), ConstantExpression(replacement)
-    )
-
-
-def array_contains(col: "ColumnOrName", value: Any) -> Column:
-    """
-    Collection function: returns null if the array is null, true if the array contains the
-    given value, and false otherwise.
-
-    Parameters
-    ----------
-    col : :class:`~pyspark.sql.Column` or str
-        name of column containing array
-    value :
-        value or column to check for in array
-
-    Returns
-    -------
-    :class:`~pyspark.sql.Column`
-        a column of Boolean type.
-
-    Examples
-    --------
-    >>> df = spark.createDataFrame([(["a", "b", "c"],), ([],)], ['data'])
-    >>> df.select(array_contains(df.data, "a")).collect()
-    [Row(array_contains(data, a)=True), Row(array_contains(data, a)=False)]
-    >>> df.select(array_contains(df.data, lit("a"))).collect()
-    [Row(array_contains(data, a)=True), Row(array_contains(data, a)=False)]
-    """
-    value = _get_expr(value)
-    return _invoke_function("array_contains", _to_column(col), value)
-
-
-def avg(col: "ColumnOrName") -> Column:
-    """
-    Aggregate function: returns the average of the values in a group.
-
-    .. versionadded:: 1.3.0
-
-    .. versionchanged:: 3.4.0
-        Supports Spark Connect.
-
-    Parameters
-    ----------
-    col : :class:`~pyspark.sql.Column` or str
-        target column to compute on.
-
-    Returns
-    -------
-    :class:`~pyspark.sql.Column`
-        the column for computed results.
-
-    Examples
-    --------
-    >>> df = spark.range(10)
-    >>> df.select(avg(col("id"))).show()
-    +-------+
-    |avg(id)|
-    +-------+
-    |    4.5|
-    +-------+
-    """
-    return _invoke_function_over_columns("avg", col)
-
-
-def sum(col: "ColumnOrName") -> Column:
-    """
-    Aggregate function: returns the sum of all values in the expression.
-
-    .. versionadded:: 1.3.0
-
-    .. versionchanged:: 3.4.0
-        Supports Spark Connect.
-
-    Parameters
-    ----------
-    col : :class:`~pyspark.sql.Column` or str
-        target column to compute on.
-
-    Returns
-    -------
-    :class:`~pyspark.sql.Column`
-        the column for computed results.
-
-    Examples
-    --------
-    >>> df = spark.range(10)
-    >>> df.select(sum(df["id"])).show()
-    +-------+
-    |sum(id)|
-    +-------+
-    |     45|
-    +-------+
-    """
-    return _invoke_function_over_columns("sum", col)
-
-
-def max(col: "ColumnOrName") -> Column:
-    """
-    Aggregate function: returns the maximum value of the expression in a group.
-
-    .. versionadded:: 1.3.0
-
-    .. versionchanged:: 3.4.0
-        Supports Spark Connect.
-
-    Parameters
-    ----------
-    col : :class:`~pyspark.sql.Column` or str
-        target column to compute on.
-
-    Returns
-    -------
-    :class:`~pyspark.sql.Column`
-        column for computed results.
-
-    Examples
-    --------
-    >>> df = spark.range(10)
-    >>> df.select(max(col("id"))).show()
-    +-------+
-    |max(id)|
-    +-------+
-    |      9|
-    +-------+
-    """
-    return _invoke_function_over_columns("max", col)
-
-
-def mean(col: "ColumnOrName") -> Column:
-    """
-    Aggregate function: returns the average of the values in a group.
-    An alias of :func:`avg`.
-
-    .. versionadded:: 1.4.0
-
-    .. versionchanged:: 3.4.0
-        Supports Spark Connect.
-
-    Parameters
-    ----------
-    col : :class:`~pyspark.sql.Column` or str
-        target column to compute on.
-
-    Returns
-    -------
-    :class:`~pyspark.sql.Column`
-        the column for computed results.
-
-    Examples
-    --------
-    >>> df = spark.range(10)
-    >>> df.select(mean(df.id)).show()
-    +-------+
-    |avg(id)|
-    +-------+
-    |    4.5|
-    +-------+
-    """
-    return _invoke_function_over_columns("mean", col)
-
-
-def min(col: "ColumnOrName") -> Column:
-    """
-    Aggregate function: returns the minimum value of the expression in a group.
-
-    .. versionadded:: 1.3.0
-
-    .. versionchanged:: 3.4.0
-        Supports Spark Connect.
-
-    Parameters
-    ----------
-    col : :class:`~pyspark.sql.Column` or str
-        target column to compute on.
-
-    Returns
-    -------
-    :class:`~pyspark.sql.Column`
-        column for computed results.
-
-    Examples
-    --------
-    >>> df = spark.range(10)
-    >>> df.select(min(df.id)).show()
-    +-------+
-    |min(id)|
-    +-------+
-    |      0|
-    +-------+
-    """
-    return _invoke_function_over_columns("min", col)
-
-
-def count(col: "ColumnOrName") -> Column:
-    """
-    Aggregate function: returns the number of items in a group.
-
-    .. versionadded:: 1.3.0
-
-    .. versionchanged:: 3.4.0
-        Supports Spark Connect.
-
-    Parameters
-    ----------
-    col : :class:`~pyspark.sql.Column` or str
-        target column to compute on.
-
-    Returns
-    -------
-    :class:`~pyspark.sql.Column`
-        column for computed results.
-
-    Examples
-    --------
-    Count by all columns (start), and by a column that does not count ``None``.
-
-    >>> df = spark.createDataFrame([(None,), ("a",), ("b",), ("c",)], schema=["alphabets"])
-    >>> df.select(count(expr("*")), count(df.alphabets)).show()
-    +--------+----------------+
-    |count(1)|count(alphabets)|
-    +--------+----------------+
-    |       4|               3|
-    +--------+----------------+
-    """
-    return _invoke_function_over_columns("count", col)
-
-
-@overload
-def transform(col: "ColumnOrName", f: Callable[[Column], Column]) -> Column:
-    ...
-
-
-@overload
-def transform(col: "ColumnOrName", f: Callable[[Column, Column], Column]) -> Column:
-    ...
-
-
-def transform(
-    col: "ColumnOrName",
-    f: Union[Callable[[Column], Column], Callable[[Column, Column], Column]],
-) -> Column:
-    """
-    Returns an array of elements after applying a transformation to each element in the input array.
-
-    .. versionadded:: 3.1.0
-
-    .. versionchanged:: 3.4.0
-        Supports Spark Connect.
-
-    Parameters
-    ----------
-    col : :class:`~pyspark.sql.Column` or str
-        name of column or expression
-    f : function
-        a function that is applied to each element of the input array.
-        Can take one of the following forms:
-
-        - Unary ``(x: Column) -> Column: ...``
-        - Binary ``(x: Column, i: Column) -> Column...``, where the second argument is
-            a 0-based index of the element.
-
-        and can use methods of :class:`~pyspark.sql.Column`, functions defined in
-        :py:mod:`pyspark.sql.functions` and Scala ``UserDefinedFunctions``.
-        Python ``UserDefinedFunctions`` are not supported
-        (`SPARK-27052 <https://issues.apache.org/jira/browse/SPARK-27052>`__).
-
-    Returns
-    -------
-    :class:`~pyspark.sql.Column`
-        a new array of transformed elements.
-
-    Examples
-    --------
-    >>> df = spark.createDataFrame([(1, [1, 2, 3, 4])], ("key", "values"))
-    >>> df.select(transform("values", lambda x: x * 2).alias("doubled")).show()
-    +------------+
-    |     doubled|
-    +------------+
-    |[2, 4, 6, 8]|
-    +------------+
-
-    >>> def alternate(x, i):
-    ...     return when(i % 2 == 0, x).otherwise(-x)
-    ...
-    >>> df.select(transform("values", alternate).alias("alternated")).show()
-    +--------------+
-    |    alternated|
-    +--------------+
-    |[1, -2, 3, -4]|
-    +--------------+
-    """
-    raise NotImplementedError
