+++ tools/pythonpkg/tests/fast/spark/test_spark_dataframe.py
@@ -1,8 +1,8 @@
 import pytest
 
-_ = pytest.importorskip("duckdb.spark")
+_ = pytest.importorskip("duckdb.experimental.spark")
 
-from duckdb.spark.sql.types import (
+from duckdb.experimental.spark.sql.types import (
     LongType,
     StructType,
     BooleanType,
@@ -14,7 +14,7 @@ from duckdb.spark.sql.types import (
     ArrayType,
     MapType,
 )
-from duckdb.spark.sql.functions import col, struct, when
+from duckdb.experimental.spark.sql.functions import col, struct, when
 import duckdb
 import re
 
@@ -91,7 +91,7 @@ class TestDataFrame(object):
         assert res == [Row(a=42, b=True), Row(a=21, b=False)]
 
     def test_df_creation_coverage(self, spark):
-        from duckdb.spark.sql.types import StructType, StructField, StringType, IntegerType
+        from duckdb.experimental.spark.sql.types import StructType, StructField, StringType, IntegerType
 
         data2 = [
             ("James", "", "Smith", "36636", "M", 3000),
@@ -192,7 +192,7 @@ class TestDataFrame(object):
         )
 
     def test_df_columns(self, spark):
-        from duckdb.spark.sql.functions import col, struct, when
+        from duckdb.experimental.spark.sql.functions import col, struct, when
 
         structureData = [
             (("James", "", "Smith"), "36636", "M", 3100),
