+++ tools/pythonpkg/tests/fast/spark/test_spark_runtime_config.py
@@ -1,6 +1,6 @@
 import pytest
 
-_ = pytest.importorskip("duckdb.spark")
+_ = pytest.importorskip("duckdb.experimental.spark")
 
 
 class TestSparkRuntimeConfig(object):
