+++ tools/pythonpkg/tests/fast/spark/test_spark_types.py
@@ -1,8 +1,8 @@
 import pytest
 
-_ = pytest.importorskip("duckdb.spark")
-from duckdb.spark.sql.types import Row
-from duckdb.spark.sql.types import (
+_ = pytest.importorskip("duckdb.experimental.spark")
+from duckdb.experimental.spark.sql.types import Row
+from duckdb.experimental.spark.sql.types import (
     StringType,
     BinaryType,
     BitstringType,
